% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_tbl.R
\name{fish_tbl}
\alias{fish_tbl}
\title{Create fish encounter tables}
\usage{
fish_tbl(.data, .reach_code, .nativity)
}
\arguments{
\item{.data}{A dataframe containing fish encounter data}

\item{.reach_code}{A field containg the reach code of the encounter record}

\item{.nativity}{Visualize "native" or "nonnative" records}
}
\value{
A kable data visualization
}
\description{
Create fish encounter tables
}
