% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_waterdata.R
\name{get_water_yr}
\alias{get_water_yr}
\title{Fetch and process waterdata for nnf standard visualizations}
\usage{
get_water_yr(.gage, .year)
}
\arguments{
\item{.gage}{Field containing the usgs station codes}

\item{.year}{The year of interest}
}
\value{
A dataframe
}
\description{
Fetch and process waterdata for nnf standard visualizations
}
