% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{hydro}
\alias{hydro}
\title{Create hydrograph with sampling days}
\usage{
hydro(.data, .gage_code, .param, .pts = TRUE)
}
\arguments{
\item{.data}{A dataframe structured to spec}

\item{.gage_code}{The USGS gage code for the desired station}

\item{.param}{The data set of interest (either mean_discharge or mean_temperature)}

\item{.pts}{Visualize sampling events as points or lines}
}
\value{
A dataframe
}
\description{
Create hydrograph with sampling days
}
