% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rch_gage.R
\name{set_rch_gage}
\alias{set_rch_gage}
\title{Get associated USGS gage codes}
\usage{
set_rch_gage(.data, .reach_code, .river_mile)
}
\arguments{
\item{.data}{A dataframe containg siet information including reach codes and river miles}

\item{.reach_code}{Field containing reach codes}

\item{.river_mile}{Field containing river miles}
}
\value{
A dataframe
}
\description{
Get associated USGS gage codes
}
