% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_rch_nnf}
\alias{tbl_rch_nnf}
\title{Reach table for NNF removal projects}
\format{
A table with 9 records and 11 fields
}
\usage{
tbl_rch_nnf
}
\description{
A dataset containing nnf removal reaches roughly defined for standard figure generation
}
\keyword{datasets}
