% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_spp_sc}
\alias{tbl_spp_sc}
\title{Complete species table}
\format{
A table with 84 records and 6 fields
}
\usage{
tbl_spp_sc
}
\description{
A dataset containing species codes from CRRP & CPW, also contains size-class grouping ids
}
\keyword{datasets}
